Implement hardware init and task logic
