+++ src/tasks/Task_Logger.cpp
@@ -1,6 +1,5 @@
 #include "common.h"
 #include "freertos/task.h"
-codex/enable-task-watchdog-and-add-handlers
 #include "esp_task_wdt.h"
 
 #include "SD_MMC.h"
@@ -87,20 +86,17 @@ static void ensureLogFiles() {
     indexFile = SD_MMC.open(indexPath, FILE_APPEND);
   }
 }
-main
 
 void Task_Logger(void *pvParameters) {
   log_blob_t blob;
   esp_task_wdt_add(NULL);
   for (;;) {
-codex/enable-task-watchdog-and-add-handlers
-    esp_task_wdt_reset();
     if (xQueueReceive(q_log, &blob, pdMS_TO_TICKS(1000))) {
-      // TODO: write blob to SD card
-
-    if (xQueueReceive(q_log, &blob, portMAX_DELAY)) {
       ensureLogFiles();
-      if (!dataFile || !indexFile) continue;
+      if (!dataFile || !indexFile) {
+        esp_task_wdt_reset();
+        continue;
+      }
 
       uint32_t offset = dataFile.size();
       uint16_t len = blob.len;
@@ -112,7 +108,7 @@ codex/enable-task-watchdog-and-add-handlers
       indexFile.flush();
 
       rotateLogs();
-main
     }
+    esp_task_wdt_reset();
   }
 }
