+++ src/tasks/Task_Uplink.cpp
@@ -10,12 +10,7 @@ void Task_Uplink(void *pvParameters) {
   event_t evt;
   esp_task_wdt_add(NULL);
   for (;;) {
-codex/enable-task-watchdog-and-add-handlers
-    esp_task_wdt_reset();
     if (xQueueReceive(q_events, &evt, pdMS_TO_TICKS(1000))) {
-      // TODO: send event via Meshtastic
-
-    if (xQueueReceive(q_events, &evt, portMAX_DELAY)) {
       const uint8_t *payload = reinterpret_cast<const uint8_t *>(&evt);
       const size_t len = sizeof(evt);
 
@@ -39,7 +34,7 @@ codex/enable-task-watchdog-and-add-handlers
         memcpy(blob.data, payload, blob.len);
         xQueueSend(q_log, &blob, 0);
       }
-main
     }
+    esp_task_wdt_reset();
   }
 }
