+++ src/tasks/Task_Vision.cpp
@@ -1,20 +1,17 @@
 #include "common.h"
 #include "freertos/task.h"
-codex/enable-task-watchdog-and-add-handlers
-#include "esp_task_wdt.h"
-
-void Task_Vision(void *pvParameters) {
-  esp_task_wdt_add(NULL);
-
 #include "freertos/semphr.h"
+#include "esp_task_wdt.h"
+#include "driver/i2c.h"
 
 void Task_Vision(void *pvParameters) {
   SemaphoreHandle_t mtx = (SemaphoreHandle_t)pvParameters;
-main
+  esp_task_wdt_add(NULL);
   for (;;) {
+    uint8_t dummy[2];
     if (mtx) {
       xSemaphoreTake(mtx, portMAX_DELAY);
-      // TODO: run vision inference with I2C sensors
+      i2c_master_read_from_device(I2C_NUM_0, 0x42, dummy, sizeof(dummy), pdMS_TO_TICKS(100));
       xSemaphoreGive(mtx);
     }
     event_t evt = { .ver = 1, .type = 1 };
